+++ src/SharpX/Contract.cs
@@ -1,0 +1,79 @@
+using System;
+using System.Linq;
+using System.Runtime.CompilerServices;
+
+namespace SharpX
+{
+    public static class Contract
+    {
+        [MethodImpl(MethodImplOptions.AggressiveInlining)]
+        public static void DisallowNull(string argumentName, object value) =>
+            _ = value ?? throw new ArgumentNullException(argumentName, $"{argumentName} cannot be null.");
+        
+        [MethodImpl(MethodImplOptions.AggressiveInlining)]
+        public static void DisallowDefault<T>(string argumentName, T value)
+            where T : struct
+        {
+            if (value.Equals(default(T))) throw new ArgumentNullException(argumentName, $"{argumentName} cannot be default.");
+        }
+
+        [MethodImpl(MethodImplOptions.AggressiveInlining)]
+        public static void DisallowWhiteSpace(string argumentName, string value)
+        {
+            if (value.Any(c => char.IsWhiteSpace(c))) throw new ArgumentException(
+                $"{argumentName} cannot be made of or contains only white spaces.", argumentName);
+        }    
+        
+        [MethodImpl(MethodImplOptions.AggressiveInlining)]
+        public static void DisallowEmptyWhiteSpace(string argumentName, string value)
+        {
+            if (value.Trim() == string.Empty) throw new ArgumentException(
+                $"{argumentName} cannot be empty or contains only white spaces.", argumentName);
+        }
+        
+        [MethodImpl(MethodImplOptions.AggressiveInlining)]
+        public static void DisallowAnyEmptyWhiteSpace(string argumentName, string[] value)
+        {
+            if (value.Any(s => s.Trim() == string.Empty)) throw new ArgumentException(
+                $"{argumentName} items cannot be empty or contains only white spaces.", argumentName);
+        }
+        
+        [MethodImpl(MethodImplOptions.AggressiveInlining)]
+        public static void DisallowNegative(string argumentName, int value)
+        {
+            if (value < 0) throw new ArgumentException(argumentName, $"{argumentName} cannot be lesser than zero.");
+        }
+        
+        [MethodImpl(MethodImplOptions.AggressiveInlining)]
+        public static void DisallowNegativeZero(string argumentName, int value)
+        {
+            if (value < 1) throw new ArgumentException(argumentName, $"{argumentName} cannot be lesser than one.");
+        }
+        
+        [MethodImpl(MethodImplOptions.AggressiveInlining)]
+        public static void DisallowOdd(string argumentName, int value)
+        {
+            if (value % 2 != 0) throw new ArgumentException(argumentName, $"{argumentName} cannot be odd.");
+        }
+
+        [MethodImpl(MethodImplOptions.AggressiveInlining)]
+        public static void RestrictArraySize<T>(string argumentName, int length, T[] value)
+        {
+            if (value.Length < length) throw new ArgumentException(
+                $"{argumentName} cannot contain less than {length} elements.", argumentName);
+        }
+        
+        [MethodImpl(MethodImplOptions.AggressiveInlining)]
+        public static void DisallowEmptyArray<T>(string argumentName, T[] value)
+        {
+            if (value.Length == 0) throw new ArgumentException($"{argumentName} cannot be empty.", argumentName);
+        }
+        
+        [MethodImpl(MethodImplOptions.AggressiveInlining)]
+        public static void AllowGuidOnly(string argumentName, string value)
+        {
+            try { Guid.Parse(value); }
+            catch { throw new ArgumentException($"{argumentName} must be a correctly formatted GUID.", argumentName); }
+        }
+    }
+}